# Further reduce Check{Decode,Encode} mem allocs.