This modifies the checksum function to return a copy the 4 checksum bytes rather than a slice to avoid escaping to the heap thereby making it zero alloc.

It also consists of a preliminary commit to add a benchmark for `CheckEncode`.

The following is a before and after comparison of a typical check encode and decode:

```
benchmark              old ns/op    new ns/op   delta
-------------------------------------------------------
BenchmarkCheckDecode   1382         1356        -1.88%
BenchmarkCheckEncode   1747         1706        -2.35%

benchmark              old allocs   new allocs  delta
-------------------------------------------------------
BenchmarkCheckDecode   2            1           -50.00%
BenchmarkCheckEncode   4            3           -25.00%

benchmark              old bytes    new bytes   delta
-------------------------------------------------------
BenchmarkCheckDecode   80           48          -40.00%
BenchmarkCheckEncode   160          128         -20.00%
```